pyproject.toml
[project]
name = "langgraph-aws-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autoevals>=0.0.129",
    "boto3>=1.39.10",
    "braintrust[otel]>=0.1.8",
    "braintrust-langchain>=0.0.2",
    "langchain-aws>=0.2.28",
    "langchain-core>=0.3.70",
    "langgraph>=0.5.4",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-instrumentation-botocore>=0.41b0",
    "strands-agents>=1.4.0",
]
